examples/2_hackernews/requirements.txt
Flask==1.0.3
Flask-DebugToolbar==0.10.1
Flask-Htmldoom>=0.1<0.2
